+++ CHANGELOG.md
@@ -18,6 +18,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Added `getResolvedParameters` method
 
+### Changed
+
+- Updated `getNamedRoutes` and `getNamedRoute` methods to replace wildcards with parameters.
+
 ## [3.0.2] - 2023.06.23
 
 ### Fixed